Allow use of alternative indexes.
Whitehall needs to be able to index things against two different
indexes, one for specialist content and one for government content.
This change should pass all requests to the default index as before, but allow any request to override the default by passing in
a different index name parameter.

With this in place we will not need the path prefix (not that I
think we were using it anyway), so it has been removed too.

N.B. The pull request to rummager to allow hosting of multiple
different backends is here: https://github.com/alphagov/rummager/pull/26