Port tests for Rummageable.index to Rummageable::Index
Rummageable::Index gives you an index object that you can use to add
single documents (with `Index#add`) or multiple documents (with
`Index#add_batch`). Call `#add` when you've just got one document to add
to the index and need the call to run quickly (such as when you're
handling an HTTP request). Call `#add_batch` when you have lots of
documents to index and can cope with a slow/blocking API call (e.g. in a
Rake task).