Change delete to take an options hash
We've belatedly realised that it would be much nicer for the new delete
endpoint to take an options parameter instead of a second string
parameter.  (There may be more options in future, and this approach
means it's clear in the calling code what the new parameter is for.)

This is a breaking change from the syntax used in 1.1.0, but no apps are
yet using that syntax.  It is not a breaking change from the syntax used
in the previous 1.0.1 version.