Version 1.2.0
Change the delete endpoint to take an options hash as its second
parameter instead of a string.

This is a breaking change from the syntax used in 1.1.0, but no apps are
yet using that syntax.  It is not a breaking change from the syntax used
in the previous 1.0.1 version.  Therefore, we're not bumping all the way
to 2.0.0.